pyproject.toml
[project]
name = "habits"
version = "0.1.0"
description = ""
authors = [
    {name = "Roman Pogoreltcev",email = "romzes26@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "requests (>=2.32.5,<3.0.0)",
    "celery (>=5.6.0,<6.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
